LP#1251394: Reingest streamlining, schema realigning, rebasing
I've rebased this to master and undertake some work to allow streamlined
reingest of specific index definitions.  The API remains essentially backwards
compatable, but is reingest_metabib_field_entries() is extended to accept a
list of index definition IDs to which it should restrict its work.  If that
optional parameter is not passed, all index definitions are processed.

This has the benefit of skipping XML transformation for index definitions
which won't produce output useful to the requested reingest process, which
should provide a significant speed boost for some situations.

The upgrade schema has now been realigned with some missing bits from the
baseline, but the baseline may still be lacking some from the upgrade.

The data upgrade script has been made to use the existing index definitions
where they match the needs of the new ones, and to use the new reindexing
options.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>