LP#1708951: fix tabbing in webstaff catalog app for Firefox
This patch works around an issue where Firefox appears to require
that the iframe for the embedded OPAC not have a null catalog_url
value. It also fixes an issue where navigating to the OPAC View tab
after retrieving the record with a different default view set would
cause it to jump back to that default view... the first time.

To test:

[1] In the web staff client, retrieve a bib record and set the default
    view to a tab other than 'OPAC View'.
[2] Retrieve another bib by TCN or ID. Note that although the default tab
    will briefly display, the page will jump over to the OPAC view and
    display an empty record.
[3] Apply the patch and repeat step 2. This time, it should display the
    record with the current default view.
[4] Select the 'OPAC View' tab and verify that it loads the record details
    page.

Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>
Signed-off-by: Beth Willis <willis@noblenet.org>
Signed-off-by: Cesar Velez <cesar.velez@equinoxinitiative.org>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>