LP#1708048: Fix ld issues by renaming libraries.
We add the lib... prefix to all C library names as expected by most
linkers.  This necessitates changes to the opensrf.xml configuration
file.

Debian 9's ld also apparently requires that ldconfig be run again
after the Evergreen C libraries are installed, so we add that step to
the installation instructions as well.

Signed-off-by: Jason Stephenson <jason@sigio.com>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>