LP#1178802: force ceiling dates to be end of day
This patch ensures that when a hard due date (or hard due date value)
is created or edited, the ceiling date is set to the end of that day.

This fixes a problem where a ceiling date would not be applied for
loans checked out earlier on the ceiling date itself.

This patch intentionally does not touch existing hard due date
values.  It also does not use the database or middle layer to coerce
the ceiling dates to end of day to leave upon the possiblity that
an Angular version of the hard due date editor will allow operators
to set both date and time. (Note that Dojo does not readily have a
combined date/time picker.)

To test
-------
[1] Apply the patch, then in the hard due date editor, verify that
    when creating a hard due date _value_, the date gets saved as
    ending at 23:59:59 in the local time zone.
[2] Similarly, verify that when editing an existing hard due date value,
    the date gets saved as ending at 23:59:59.
[3] Repeat steps 1 and 2, but this time for hard due dates.

Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>
Signed-off-by: Michele Morgan <mmorgan@noblenet.org>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>