LP#1700635: enable eg-marc-editor for queued records in MARC Batch Import/Export
This enables the angular modal marc-editor to be called from
the vandelay dojo UI in the iFrame.

To test:

[1] In the web staff client, go to MARC Batch Import/Export and load
    some records in a Vandelay queue.
[2] Inspect the queue and choose View MARC.
[3] Click the Edit button, and verify that it brings up the
    web staff MARC editor.
[4] Make a change, then click Modify, then click Use Edits.
[5] Verify that the MARC view reflects the updated record.
[6] Go back to the queue. If the bib was /not/ imported into
    Evergreen, the bib attributes may change.
[7] Verify that steps 1-6 in the XUL client work (and that the XUL
    MARC editor is displayed).

Note that this patch doesn't ensure that the modal can save the
record and be dismissed via just a single click of the Modify button;
an issue was discovered with that that could not be resolved in time
for beta2.

Signed-off by: Cesar Velez <cesar.velez@equinoxinitiative.org>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>