LP#1717777: absorb all PO files for a locale
We assumed only one file per locale, but apache <Location>s because are not
independent WRT locale helpers in mod_perl, we need to load all files from
all <Location> directives for a given locale.

NOTE: there is a risk of translation collision, where the same key string in
two PO files have different translations, but in our context that is almost
surely a mistake.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Ben Shum <ben@evergreener.net>