LP#1718301: catch it when multiple connection attempts fail
Lovefield only allows a single connection at a time, and throws
an exception if a second attempt is made.  In the context of a
page app, wrapping the connection attempt in a promise takes
care of it, but for the test suite, the egLoveField service
gets completely reinitialized... but Lovefield itself still
thinks it's in the middle of connecting.

This patch catches the exception and allows the test suite to
pass, without otherwise affecting use in a real browser.

Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>
Signed-off-by: Ben Shum <ben@evergreener.net>