LP#1378383-WebStaff Fix Item circ history to obey max setting
The circ.item_checkout_history.max lib setting should be respected
in both Circ History tabs, and also check for the permission to
VIEW_COPY_CHECKOUT_HISTORY for the current user.

Also, There was ambiguity for the user to tell the difference
(from the displayed message) if an item truly has no circulations,
or if they see no circs due to the relevant permission setting
being set to 0 (Maximum previous checkouts.) This changes the templates
a bit to make that clearer.

Signed-off by: Cesar Velez <cesar.velez@equinoxinitiative.org>

Signed-off-by: Jason Etheridge <jason@EquinoxInitiative.org>