LP#1643927 Copy note captures initials; shows date
Fixes thinko in copy note staff initials tracking.  Offloads the
formatting of the copy note to egStrings.  Adds the workstation org unit
to the initials block a la the XUL client.

This does not add the create date to the initials block as with the XUL
client, because the create date is part of the note.  Instead, we now
display the note create date in the copy note display form.

Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Cesar Velez <cesar.velez@equinoxinitiative.org>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>