LP 1768715: Use named arguments in function calls in pingest.pl.
This makes it work, more or less, with Evergreen 3.0 and the changes
to metabib.reingest_field_entries.  Note that the new skip_display
parameter is not set, so you might get collisions between the browse
ingest and one of the other ingest processes if they both work on the
same record at once.

I will make a branch for use with Evergreen 3.0 until I upgrade to
Evergreen 3.0 and merge that branch into master.

Signed-off-by: Jason Stephenson <jason@sigio.com>
Signed-off-by: Bill Erickson <berickxx@gmail.com>