LP#1689325 - require most modals have explicit 'exit' or 'cancel' action inside the modal
Disable the default bahavior of bootstrap modals that lets
the user click outside the modal to dismiss or cancel it.
Similar behavior is not in the XUL, and this will make
unintentional clicks outside the modal less prone to losing
work in said modal.

Signed-off by: Cesar Velez <cesar.velez@equinoxinitiative.org>

Signed-off-by: Mike Rylander <mrylander@gmail.com>