LP#712490 Vandelay merge-based field replacement
Modify the vandelay field replacement logic to merge replacement data
into existing field data where possible, instead of simply deleting then
re-adding the data.  The key benefit is that subfields retain their order
in affected MARC fields.  As a result, for example, propagating
authority field changes to bib records will no longer result in the $0
field being pushed to the front of the field when it was previously that
last subfield.

Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Jane Sandberg <sandbej@linnbenton.edu>
Signed-off-by: Jason Stephenson <jason@sigio.com>