LP#1727487 Webstaff display fields collapseMulti option
Support an option in the webstaff-side display field munging code to
collapse array/multi values down to a single comma-separated string.

This is useful for buidling displays (grids especially) where you have a
single spot to put a field's value (e.g. a list of ISBN's) and don't want
to munge the data by hand in each UI.

Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Dan Wells <dbw2@calvin.edu>