LP#1723977: Move no-LURIs test to be a peer of no-copies test
The correct place to test the no-LURIs condition is as a peer to the
no-copies test, when searching in staff mode.  We therefore move that
test with this commit.

In addition, the bib attribute vector may be null, in the case of a
record with no bib source or LURI.  When testing emptiness we need to
ask if the field IS NULL rather than just testing its contents.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Jason Boyer <JBoyer@library.in.gov>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>