LP#1734737: Display originating template UI
Until the (planned for 3.1) ability to clone XUL-created reports in the web
staff client is completed, it would be useful to at least display to the user
the origin UI in which a template was created. This will allow staff to, for
the immediate term, decide which UI they should use to deal with a particular
template, re cloning etc.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Chris Sharp <csharp@georgialibraries.org>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>