Lp 1350916: marc_export --uris option release notes.
Add release notes for the new --uris option on marc_export.

To test the new option, run marc_export to export records with
holdings for a given library using the --library and --items options.
If that library also has electronic resources, do another export with
the same options but add the --uris option and direct this output to a
different file.  Compare the file sizes.  The second file should be
larger than the first.

Signed-off-by: Jason Stephenson <jason@sigio.com>
Signed-off-by: Dan Wells <dbw2@calvin.edu>