LP#1729922: correctly display most recent transit in item status
This patch ensures that the 'Most Recent Transit' display in
the web staff client's item status page shows the latest transit,
not the oldest one. This also fixes a bug where the transit
receive time never got displayed.  As a side effect, this patch
also ensures that only the transit in question is fetched from
the database, not all transits on the item.

To test
-------
[1] Set up an item with some closed transits and display in in
    detail view in the Item Status page. Note that the
    /oldest/ transit is displayed and that the transit
    receive time field is blank.
[2] Apply the patch and repeat step 1.
[3] This time, the most recent transit should be displayed, and
    if that transit was received, the receive time should be
    displayed.

Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>
Signed-off-by: Chris Sharp <csharp@georgialibraries.org>
Signed-off-by: Jason Etheridge <jason@EquinoxInitiative.org>