LP#1730470 Restore XUL serial receive compatibility
The older XUL interface does not have quite the same fleshing as the
new web-staff, so adding the 'move previous' code into the unitize
function breaks the old interface.

Since we already are looking up the distribution to get the setting,
let's just tack it on when it is missing.

To test
-------
[1] Set up a serial subscription with predicted issuances, then from
    the XUL serials control view attempt to receive an issue. Note
    that there will be a crash.
[2] Apply the patch and repeat step #1; this time, the issuance
    receive should work.
[3] Verify that receiving issues via the web staff interface continues
    to work.

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>
Signed-off-by: Mike Rylander <mrylander@gmail.com>