LP#1739669 - add PaymentType and persistkey bill payment hist grid
This adds "circ.patron.billhistory_payments" as the persist-key for that
grid and also adds a Payment Type column. Note that the Payment Type is
that of the individual payment, while the Last Payment Type is that
of the most recent payment on the billing in question.

To test
-------
[1] Set up a patron bill that has multiple payments on it off
    different types (e.g., cash, credit, etc.)
[2] Apply the patch.
[3] In the payments tab of the Bill History page in the web staff
    client, note that there is now a Payment Type column available.
    Verify that it matches the types of the payments entered in
    step 1.
[4] Verify that column settings for that payment history grid can be
    saved.

Signed-off by: Cesar Velez <cesar.velez@equinoxinitiative.org>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>

Signed-off-by: Jason Etheridge <jason@EquinoxInitiative.org>