LP#1724052: move stat-cat cache initialization to patron search service
When the patron search was split out of the patron app per bug 1701001,
the startup code that copied the current patron stat-cat list into
egCore.env was not moved over as well and remained in the patron app. As
a consequence, the patron search-to-hold dialog would fail to display
any patrons in the results list that had statistical categories assigned.
This patch moves that bit of initialization into the service.

To test
-------
[1] Ensure that some patron records have statistical categories assigned.
[2] In the web staff client, perform a catalog search, place a hold,
    then use the 'Patron Search' button. Perform a search; note that
    patrons with statistical categories are not displayed and that
    errors are present in the browser debug console.
[3] Apply the patch and repeat step #2. This time, patrons that have
    stat cats should be included in the results.
[4] Perform a patron search in the main patron search interface;
    verify that patrons with statistical categories are retrieved.
[5] Directly go to the page listing stat cats for a patron (e.g.,
    /eg/staff/circ/patron/$ID/stat_cats). Verify that the name
    of the OU that owns the stat cat is displayed.

Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>
Signed-off-by: Cesar Velez <cesar.velez@equinoxinitiative.org>
Signed-off-by: Bill Erickson <berickxx@gmail.com>