LP#1743639: opac_visible on acplg is not what it seems
The asset.copy_location_group table has a column called opac_visible.
However, unlike other tables with a column of that name, it controls the
visibility of the location group itself, not the visibility of the copies
within the group.  This is regretable because the 3.0+ search code thinks
that it should hide copies.

This commit adjusts the default location group filter applied to patron
searches so that they act as expected.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>