LP#1743639: Test location as proxy for location group
Location groups are a table remove from copies, and can cause very large
updates to the copy vis cache.  So, we will expand location groups to
locations, in situ, to avoid this problem and the need to supply a trigger
to update th copy vis cache for groups.  This will allow us to reclaim
the location group bit in the visibility cache namespace if necessary, at
a later date.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>