LP#1738249 - Webstaff CircLib should point to item's circ_lib plus misc
This addresses some display issues in the webstaff Item Status grid.
Besides, correcting the Circ Lib to be the item's and not the
circulation,this also fleshes the owning lib. Additionally, I've applied
miker's fix to 1743819 in order to prevent a regression.

Signed-off by: Cesar Velez <cesar.velez@equinoxinitiative.org>
Signed-off-by: Terran McCanna <tmccanna@georgialibraries.org>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>
Signed-off-by: Jason Stephenson <jason@sigio.com>