LP#1745001 - action_trigger_runner.pl max-sleep argument broken.
Fix the --max-sleep argument so it accepts an integer number of max
seconds to sleep waiting for a lock file to disappear.

Testing:

Before Fix: Add --max-sleep=60 to action_trigger_runner.pl and view the
error message that results.

After Fix: Run the same command and notice that it no longer errors out.

Signed-off-by: Josh Stompro <stomproj@larl.org>
Signed-off-by: Cesar Velez <cesar.velez@equinoxinitiative.org>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>