LP#1745610 - Deprecate original hold_targeter.pl
With the new hold targeter battle-tested for over a year,
it's now time to promote it to default, deprecating the
older version.  This commit renames the old targeter to
"hold_targeter_legacy.pl" and the new to "hold_targeter.pl",
and adapts the example crontab to use the new targeter's
parameter syntax.

Signed-off-by: Chris Sharp <csharp@georgialibraries.org>
Signed-off-by: Bill Erickson <berickxx@gmail.com>