LP#1635386: Improve contrast for colors used in Bills interface
The contrast between cell contents and background colors did not
meet WCAG standards. This commit provides adequate contrast to
meet AA level of WCAG standards. The red and dark red background
colors are similar to what we saw in the XUL client. The orange
for long overdue is much lighter.

Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Dawn Dale <ddale@georgialibraries.org>