LP#1746584: Adjust bib visibility test for patron browse
With updates to address visibility testing issues for LURIs, a change was made
to allow the default bib tests to supply the most appropriate query_int
operator with which to join those to non-default tests.  The browse code,
however, is all database-level and was not adjusted with the Perl code that
implements general search.  This commit addresses that issue by acknowledging
that the bib vis testing code to provide its own operator, either | or &, as
appropriate to the actual default test values.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Jason Stephenson <jason@sigio.com>